Spike on selecting cards. Check when down arrow skips a section, and if that section was a card, drop the cursor and instead highlight the card.
This is a spike to see if this method is an improvement over trying to
determine the cursor position in the section and check whether it is on
the last line at the time that the down arrow is pressed.

On Chrome you can see the cursor flutter in the section below the card,
so this method is not ideal.